chore: add cert validator for static site
