+++ internal/pkg/deploy/upload/customresource/customresource_test.go
@@ -96,8 +96,8 @@ func TestLBWS(t *testing.T) {
 			"custom-resources/wkld-custom-domain.js": {
 				Buffer: bytes.NewBufferString("service-level custom domain"),
 			},
-			"custom-resources/nlb-cert-validator.js": {
-				Buffer: bytes.NewBufferString("nlb cert"),
+			"custom-resources/wkld-cert-validator.js": {
+				Buffer: bytes.NewBufferString("service-level cert"),
 			},
 		},
 	}
@@ -106,7 +106,7 @@ func TestLBWS(t *testing.T) {
 		"EnvControllerFunction":       "manual/scripts/custom-resources/envcontrollerfunction/72297cacaeab3a267e371c17ea3f0235905b0da51410eb31c10f7c66ba944044.zip",
 		"RulePriorityFunction":        "manual/scripts/custom-resources/rulepriorityfunction/1385d258950a50faf4b5cd7deeecbc4bcc79a0d41d631e3977cffa0332e6f0c6.zip",
 		"NLBCustomDomainFunction":     "manual/scripts/custom-resources/nlbcustomdomainfunction/ac1c96e7f0823f3167b4e74c8b286ffe8f9d43279dc232d9478837327e57905e.zip",
-		"NLBCertValidatorFunction":    "manual/scripts/custom-resources/nlbcertvalidatorfunction/3b9f56301b50779e09a3495a6d7eadc42b4401f265d4cfb359543c1ad3f21769.zip",
+		"NLBCertValidatorFunction":    "manual/scripts/custom-resources/nlbcertvalidatorfunction/41aeafc64f18f82c452432a214ae83d8c8de4aba2d5df6a752b7e9a2c86833f1.zip",
 	}
 
 	// WHEN
@@ -261,14 +261,18 @@ func TestStaticSite(t *testing.T) {
 				Buffer: bytes.NewBufferString("copy assets"),
 			},
 			"custom-resources/wkld-custom-domain.js": {
-				Buffer: bytes.NewBufferString("add A record for CloudFront"),
+				Buffer: bytes.NewBufferString("service-level custom domain"),
+			},
+			"custom-resources/wkld-cert-validator.js": {
+				Buffer: bytes.NewBufferString("service-level cert"),
 			},
 		},
 	}
 	fakePaths := map[string]string{
-		"TriggerStateMachineFunction": "manual/scripts/custom-resources/triggerstatemachinefunction/edfa40b595a5a4a6d24bfb7ad6e173560a29b7d720651ccc9c87eda76b93c7dd.zip",
-		"CopyAssetsFunction":          "manual/scripts/custom-resources/copyassetsfunction/b9fc2f284cfb699b7c63efaac618d7678372792aa020141fa91053092977976d.zip",
-		"CustomDomainFunction":        "manual/scripts/custom-resources/customdomainfunction/c8c78ed9e73964c2facfa0bceb5ed6227172fd98057c53c137296fed81975672.zip",
+		"TriggerStateMachineFunction":   "manual/scripts/custom-resources/triggerstatemachinefunction/edfa40b595a5a4a6d24bfb7ad6e173560a29b7d720651ccc9c87eda76b93c7dd.zip",
+		"CopyAssetsFunction":            "manual/scripts/custom-resources/copyassetsfunction/b9fc2f284cfb699b7c63efaac618d7678372792aa020141fa91053092977976d.zip",
+		"CustomDomainFunction":          "manual/scripts/custom-resources/customdomainfunction/ac1c96e7f0823f3167b4e74c8b286ffe8f9d43279dc232d9478837327e57905e.zip",
+		"CertificateValidationFunction": "manual/scripts/custom-resources/certificatevalidationfunction/41aeafc64f18f82c452432a214ae83d8c8de4aba2d5df6a752b7e9a2c86833f1.zip",
 	}
 
 	// WHEN
@@ -276,14 +280,14 @@ func TestStaticSite(t *testing.T) {
 
 	// THEN
 	require.NoError(t, err)
-	require.Equal(t, fakeFS.matchCount, 3, "expected path calls do not match")
+	require.Equal(t, fakeFS.matchCount, 4, "expected path calls do not match")
 
 	actualFnNames := make([]string, len(crs))
 	for i, cr := range crs {
 		actualFnNames[i] = cr.Name()
 	}
 	require.ElementsMatch(t,
-		[]string{"TriggerStateMachineFunction", "CopyAssetsFunction", "CustomDomainFunction"},
+		[]string{"TriggerStateMachineFunction", "CopyAssetsFunction", "CustomDomainFunction", "CertificateValidationFunction"},
 		actualFnNames, "function names must match")
 
 	// ensure the zip files contain an index.js file.
