+++ internal/pkg/template/templates/workloads/services/static-site/cf.yml
@@ -92,6 +92,9 @@ Resources:
     Type: AWS::CloudFront::Distribution
     Properties:
       DistributionConfig:
+        {{- if .StaticSiteAlias}}
+        Aliases: [{{.StaticSiteAlias}}]
+        {{- end}}
         DefaultCacheBehavior:
           AllowedMethods: ["GET", "HEAD", "OPTIONS", "PUT", "PATCH", "POST", "DELETE"]
           ViewerProtocolPolicy: allow-all
@@ -107,6 +110,13 @@ Resources:
             # required when the origin is an S3 bucket.
             S3OriginConfig:
               OriginAccessIdentity: '' 
+        ViewerCertificate:
+          !If
+            - HasAssociatedDomain
+            - AcmCertificateArn: !Ref CertificateValidatorAction
+              MinimumProtocolVersion: TLSv1
+              SslSupportMethod:  sni-only
+            - !Ref AWS::NoValue
 
   TriggerStateMachineFunction:
     Metadata:
@@ -375,6 +385,108 @@ Resources:
                   - "route53:Describe*"
                   - "route53:ListResourceRecordSets"
                   - "route53:ListHostedZonesByName"
+                Resource: "*"
+      ManagedPolicyArns:
+        - !Sub arn:${AWS::Partition}:iam::aws:policy/service-role/AWSLambdaBasicExecutionRole
+
+  CertificateValidatorAction:
+    Metadata:
+      'aws:copilot:description': "Request and validate the certificate for your Static Site"
+    Type: Custom::CertificateValidationFunction
+    Condition: HasAssociatedDomain
+    Properties:
+      ServiceToken: !GetAtt CertificateValidationFunction.Arn
+      EnvHostedZoneId:
+        Fn::ImportValue:
+          !Sub "${AppName}-${EnvName}-HostedZone"
+      EnvName: !Ref EnvName
+      AppName: !Ref AppName
+      ServiceName: !Ref WorkloadName
+      RootDNSRole: {{ .AppDNSDelegationRole }}
+      DomainName:  {{ .AppDNSName }}
+      IsCloudFrontCertificate: true
+      Aliases: {{ if .StaticSiteAlias }} [{{ quote .StaticSiteAlias }}] {{ else }} [] {{ end }}
+
+  CertificateValidationFunction:
+    Type: AWS::Lambda::Function
+    Condition: HasAssociatedDomain
+    Properties:
+      {{- with $cr := index .CustomResources "CertificateValidationFunction" }}
+      Code:
+        S3Bucket: {{$cr.Bucket}}
+        S3Key: {{$cr.Key}}
+      {{- end }}
+      Handler: "index.handler"
+      Timeout: 900
+      MemorySize: 512
+      Role: !GetAtt 'CertificateValidatorRole.Arn'
+      Runtime: nodejs16.x
+
+  CertificateValidatorRole:
+    Metadata:
+      'aws:copilot:description': "An IAM role {{- if .PermissionsBoundary}} with permissions boundary {{.PermissionsBoundary}} {{- end}} to request and validate a certificate for your service"
+    Type: AWS::IAM::Role
+    Properties:
+      AssumeRolePolicyDocument:
+        Version: '2012-10-17'
+        Statement:
+          -
+            Effect: Allow
+            Principal:
+              Service:
+                - lambda.amazonaws.com
+            Action:
+              - sts:AssumeRole
+      {{- if .PermissionsBoundary}}
+      PermissionsBoundary: !Sub 'arn:${AWS::Partition}:iam::${AWS::AccountId}:policy/{{.PermissionsBoundary}}'
+      {{- end}}
+      Path: /
+      Policies:
+        - PolicyName: "CertValidatorPolicy"
+          PolicyDocument:
+            Version: '2012-10-17'
+            Statement:
+              - Sid: AllowAssumeRole
+                Effect: Allow
+                Action: sts:AssumeRole
+                Resource: "*"
+              - Sid: HostedZoneUpdateAndWait
+                Effect: Allow
+                Action: route53:ChangeResourceRecordSets
+                Resource: "*"
+              - Sid: HostedZoneRead
+                Effect: Allow
+                Action:
+                  - route53:ListResourceRecordSets
+                  - route53:GetChange
+                Resource: "*"
+              - Sid: ServiceCertificateDelete
+                Effect: Allow
+                Action: acm:DeleteCertificate
+                Resource: "*"
+                Condition:
+                  StringEquals:
+                    'aws:ResourceTag/copilot-application': !Sub '${AppName}'
+                    'aws:ResourceTag/copilot-environment': !Sub '${EnvName}'
+                    'aws:ResourceTag/copilot-service': !Sub '${WorkloadName}'
+              - Sid: TaggedResourcesRead
+                Effect: Allow
+                Action: tag:GetResources
+                Resource: "*"
+              - Sid: ServiceCertificateCreate
+                Effect: Allow
+                Action:
+                  - acm:RequestCertificate
+                  - acm:AddTagsToCertificate
+                Resource: "*"
+                Condition:
+                  StringEquals:
+                    'aws:ResourceTag/copilot-application': !Sub '${AppName}'
+                    'aws:ResourceTag/copilot-environment': !Sub '${EnvName}'
+                    'aws:ResourceTag/copilot-service': !Sub '${WorkloadName}'
+              - Sid: CertificateRead
+                Effect: Allow
+                Action: acm:DescribeCertificate
                 Resource: "*"
       ManagedPolicyArns:
         - !Sub arn:${AWS::Partition}:iam::aws:policy/service-role/AWSLambdaBasicExecutionRole
